Convert prepare_data and scatter_plot to entry_points
